refactor:  remove guards file
