--- packages/expo-styling/src/runtime/native/guards.ts
@@ -1,17 +1,0 @@
-import { RuntimeValue } from "../../types";
-
-export function isRuntimeValue(value: unknown): value is RuntimeValue {
-  if (!value) {
-    return false;
-  } else if (Array.isArray(value)) {
-    return value.some((v) => isRuntimeValue(v));
-  } else if (typeof value === "object") {
-    if ("type" in value && value.type === "runtime") {
-      return true;
-    } else {
-      return Object.values(value).some((v) => isRuntimeValue(v));
-    }
-  } else {
-    return false;
-  }
-}
