+++ packages/expo-styling/src/runtime/native/utils.ts
@@ -1,4 +1,4 @@
-import { SignalLike } from "../../types";
+import { RuntimeValue, SignalLike } from "../../types";
 
 export function unwrap<T>(value: T | SignalLike<T>): T {
   return value && typeof value === "object" && "get" in value
@@ -8,4 +8,20 @@ export function unwrap<T>(value: T | SignalLike<T>): T {
 
 export function kebabToCamelCase(str: string) {
   return str.replace(/-./g, (x) => x[1].toUpperCase());
+}
+
+export function isRuntimeValue(value: unknown): value is RuntimeValue {
+  if (!value) {
+    return false;
+  } else if (Array.isArray(value)) {
+    return value.some((v) => isRuntimeValue(v));
+  } else if (typeof value === "object") {
+    if ("type" in value && value.type === "runtime") {
+      return true;
+    } else {
+      return Object.values(value).some((v) => isRuntimeValue(v));
+    }
+  } else {
+    return false;
+  }
 }
